Enabled use of E component
Previously, the SEC calculations only considered the dbn_geo component of the magnetic field. Now, the algorithm accepts a Z vector that is twice as long (because it includes both components of dB as described in Amm & Viljanen 1999), and creates a transfer matrix T that is twice as long in a single dimension to accommodate this.